backend/requirements-fastapi.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==1.4.54
psycopg2-binary==2.9.1
pydantic==2.10.6
python-dotenv>=0.19.0
pandas==2.2.3
scikit-learn==1.6.1